+++ django_echarts/entities/__init__.py
@@ -2,7 +2,7 @@ from .articles import (
     ChartInfo, ChartInfoManagerMixin, LocalChartInfoManager
 )
 from .base import DwString
-from .chart_widgets import (NamedCharts, WidgetGetterMixin, merge_js_dependencies)
+from .chart_widgets import (NamedCharts, WidgetGetterMixin)
 from .containers import WidgetCollection
 from .html_widgets import (
     LinkItem, Menu, Jumbotron, Nav, Copyright, Message, ValuesPanel