+++ tests/test_interface.py
@@ -4,7 +4,7 @@ import unittest
 
 from unittest.mock import MagicMock
 
-from django_echarts.entities.chart_widgets import merge_js_dependencies
+from django_echarts.entities.widgettools import get_js_dependencies
 
 
 class MockCharts(object):
@@ -25,57 +25,57 @@ class JsMergeTestCase(unittest.TestCase):
         # One chart or one page
         self.assertListEqual(
             ['echarts'],
-            merge_js_dependencies(self.BASE_CHART)
+            get_js_dependencies(self.BASE_CHART)
         )
         self.assertListEqual(
             ['echarts', 'fujian'],
-            merge_js_dependencies(self.MAP_CHART)
+            get_js_dependencies(self.MAP_CHART)
         )
 
     def test_multiple_charts_and_pages(self):
         # Multiple charts
         self.assertListEqual(
             ['echarts', 'fujian'],
-            merge_js_dependencies(
+            get_js_dependencies(
                 self.BASE_CHART,
                 self.MAP_CHART
             )
         )
         self.assertListEqual(
             ['echarts', 'echartsgl'],
-            merge_js_dependencies(self.BASE_CHART, self.THREE_D_CHART)
+            get_js_dependencies(self.BASE_CHART, self.THREE_D_CHART)
         )
         self.assertListEqual(
             ['echarts', 'echartsgl', 'fujian'],
-            merge_js_dependencies(self.MAP_CHART, self.THREE_D_CHART)
+            get_js_dependencies(self.MAP_CHART, self.THREE_D_CHART)
         )
         self.assertListEqual(
             ['echarts', 'echartsgl', 'fujian'],
-            merge_js_dependencies(self.BASE_CHART, self.MAP_CHART, self.THREE_D_CHART)
+            get_js_dependencies([self.BASE_CHART, self.MAP_CHART, self.THREE_D_CHART])
         )
 
     def test_string_dependencies(self):
         self.assertListEqual(
             ['echarts', 'echartsgl'],
-            merge_js_dependencies('echartsgl', self.BASE_CHART)
+            get_js_dependencies('echartsgl', self.BASE_CHART)
         )
         self.assertListEqual(
             ['echarts', 'fujian'],
-            merge_js_dependencies('echarts', ['echarts', 'fujian'])
+            get_js_dependencies('echarts', ['echarts', 'fujian'])
         )
 
     def test_nested_js_dependencies(self):
         self.assertListEqual(
             ['echarts', 'echartsgl', 'fujian'],
-            merge_js_dependencies(
-                merge_js_dependencies(self.BASE_CHART), self.MAP_CHART, self.THREE_D_CHART
+            get_js_dependencies(
+                [get_js_dependencies(self.BASE_CHART), self.MAP_CHART, self.THREE_D_CHART]
             )
         )
         self.assertListEqual(
             ['echarts', 'echartsgl', 'fujian', 'zhejiang'],
-            merge_js_dependencies(
-                merge_js_dependencies(self.BASE_CHART, self.MAP_CHART),
+            get_js_dependencies([
+                get_js_dependencies(self.BASE_CHART, self.MAP_CHART),
                 self.THREE_D_CHART,
                 'zhejiang'
-            )
+            ])
         )