+++ docs/changelog.md
@@ -3,7 +3,7 @@
 ### v0.5.1 - 
 
 - 新增统一的 `dw_widget` 标签函数，原有 `dw_table` / `dw_values_panel` / `dw_link` 不再推荐使用
-- 优化js依赖项解析逻辑
+- 新增 `get_js_dependencies` 替代 `merge_js_dependencies`
 
 ## v0.5.0 - (20220318)